update nan_test to check against pandas dataframe
